Fix a segfault that can happen when creating garbage blocks
This pesky bug has been around for a long time. When creating garbage block
we might not get to create as many as we wanted because the row was already
occupied by other blocks. We need to always check for this situation, not
only when we weren't able to place a block.